requirements.txt
python-telegram-bot==20.7
requests==2.31.0
selenium==4.15.2
webdriver-manager==4.0.1